requirements.txt
fastapi==0.103.0
uvicorn==0.23.2
sqlalchemy==2.0.20
pydantic==2.4.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pytest==7.4.0
httpx==0.24.1